+++ R/family_ColumnDotPlot.R
@@ -27,8 +27,10 @@
 #' \item \code{SizeByColumnData}, a string specifying the \code{\link{colData}} field for controlling point size,
 #' if \code{SizeBy="Column data"} (see the \linkS4class{Panel} class).
 #' The specified field should contain continuous values; defaults to the first such valid field.
+#' \item \code{TooltipColumnData}, a character vector specifying \code{\link{colData}} fields to show in the tooltip.
+#' Defaults to `character(0)`, which displays only the `colnames` value of the data point.
 #' }
-#'
+#' 
 #' In addition, this class inherits all slots from its parent \linkS4class{DotPlot} and \linkS4class{Panel} classes.
 #'
 #' @section Supported methods:
@@ -156,6 +158,8 @@ setMethod("initialize", "ColumnDotPlot", function(.Object, ..., SelectionEffect=
     if (!is.null(SelectionColor)) {
         .Deprecated(msg="'SelectionColor=' is deprecated and will be ignored")
     }
+    
+    args <- .emptyDefault(args, .tooltipColData, getPanelDefault(.tooltipColData))
 
     do.call(callNextMethod, c(list(.Object), args))
 })
@@ -266,6 +270,7 @@ setMethod(".refineParameters", "ColumnDotPlot", function(x, se) {
 
     available <- cdp_cached$valid.colData.names
     x <- .replaceMissingWithFirst(x, .colorByColData, available)
+    x <- .removeInvalidChoices(x, .tooltipColData, available)
 
     assays <- dp_cached$valid.assay.names
     x <- .replaceMissingWithFirst(x, .colorByFeatNameAssay, assays)
@@ -588,6 +593,26 @@ setMethod(".colorDotPlot", "ColumnDotPlot", function(x, colorby, x_aes="X", y_ae
 
     } else {
         .colorByNoneDotPlotScale(x)
+    }
+})
+
+###############################################################
+# Tooltip
+
+setMethod(".getTooltipUI", "ColumnDotPlot", function(x, se, name) {
+    if (length(x[[.tooltipColData]]) > 0) {
+        # as.data.frame sometimes needed to fix names of items in vector
+        info <- as.data.frame(colData(se)[name, x[[.tooltipColData]], drop=FALSE])
+        info <- sapply(info, function(x) as.character(x))
+        ui <- HTML(
+            paste0(c(
+                sprintf("<strong>%s</strong>", name),
+                sprintf("%s: %s", names(info), info)
+                ), collapse = "<br />")
+            )
+        ui
+    } else {
+        name
     }
 })