+++ NEWS.md
@@ -1,3 +1,8 @@
+# iSEE 2.9.10
+
+* Enable customisation of hovering tooltip in `DotPlot` panels
+  including `colData` or `rowData` information.
+
 # iSEE 2.9.9
 
 * Allow screenshots in vignettes to use full width of pkgdown site.